separate model and computation in HMM matching
